pull most of UR into runner
